Update README.md
As I promised in our last meeting, I included all the material I have found that is useful for us. The most important parts:
1 -- Even for approximate query processing with sketches there is work done and we can integrate it to Gradoop++ after making sure it works well. No need for you to re-implement anything Annemarie. Check the Approximate Query Library!
2 -- For the experiments we can compare with GraphX which also is streaming as it sits on top of Spark Streaming API. And for a proper comparison we may use GraphTides framework.

So, lets get our Gradoop++ base functionality up and running and then we are ready to integrate an approximate library to maintain the whole stream graph approximately and precisely inside the prescribed window. :)